Cross compiling the launcher
In order to automate the builds and tests of WSL distros, we need to
be able to build the application using linux machines. This can be
done using mingw-w64.

This commit adds the autotools scripts to build the project. It also
extracts a few distro dependent variables into generated files to
handle several distros with the same code more easily.